Change Composer into an interface
The implied parameter to a Composable function is now
and interface instead of a concrete type to allow non-
breaking iteration on the implementation of the
composer.

Test: ComposerTests
Fixes: b/169406779

Relnote: """The API the Compose compiler plugin targets
has been refactored to use an interface instead of a
concrete class. The interface also no longer uses
a type parameter.

This is an internal change that should not effect source
code compatibility but is a binary breaking change."""

Change-Id: I3b9229969aa70138bc57f5e8498602f5b2dba1e6