allow to specify size of surface and take screenshot within ComposeUiTest (#310)
## Proposed Changes

Currently, screenshots can be taken using `ImageComposeScene`. However, using this approach it is not possible to interact with the UI in an automated way before taking the screenshots. On the other hand, Interaction with UI nodes is possible in testing environment (using `runComposeUiTest()`), but it uses a hard-coded size for the scene and the skia surface is not exposed to the user as it is marked as `private`.

This PR proposes to:
  - allow the user to specify custom width and height for the compose scene used during the test, by calling `runDesktopComposeUiTest(width, height)` (this mirrors the behavior on Android where you can pass custom activities using `runAndroidComposeUiTest(activityClass)`)
  - add a function to take a screenshot of the current Compose scene

## Testing

Test: Unfortunately I could not test the changes since I was unable to get the huge repository to compile using IntelliJ (any hints are very appreciated), but I did test the same functionality before using this kind of code:

```kotlin
// needs to use hard-coded size of image to work as expected (to match hard-coded size inside [ComposeUITest.desktop])
private val surface = Surface.makeRasterN32Premul(1024, 768)

@OptIn(InternalTestApi::class)
private fun DesktopComposeTestRule.takeScreenshot(): Image {
    surface.canvas.clear(Color.TRANSPARENT)
    scene.render(surface.canvas, mainClock.currentTime)
    return surface.makeImageSnapshot()
}
```

---

CLA is signed