Improve allocations during pointer handling
The previous implementation of pointer handling used a HashMap
to hold pointer information. HashMap uses Objects for keys, so
the value class for PointerId (basically an alias for a Long primitive
value) was constantly being auto-boxed because it was used as a key.

The new approach uses the leaner LongSparseArray, which was created
specifically to optimze for this use case; using primitive Long values
as keys for an array.

Bug: 283871952
Test: PointerInputTest passes, including new pointerIdArrayTest
Change-Id: I77ae2d3de1f2604b7d61ed7a1d7215815dbbf999