[Tooltip] Adding focusable parameter to both tooltip APIs
Adding focusable parameter to PlainTooltipBox and RichTooltipBox. The current tooltip is set to be focusable to have accessibility focus automatically move to the tooltip when shown, but it caused an undesirable side effect of consuming touch events while shown. This fix allows developers to set focusable to false so that the tooltip no longer consumes touch events but accessibility focus doesn't automatically move to the first element of the tooltip when shown. Due to framework limitations, we can either have accessibility focus move automatically to the tooltip OR the tooltip doesn't consume touch events BUT not both. Adding documentation about what happens when this parameter takes true or false.

Bug: 285333902
Test: N/A
Relnote: Adding focusable parameter to both tooltip APIs, so that developers can have the tooltip consume touch events or not. Be cautious that this might break accessibility focus automatic traversal.
Change-Id: Ie32d8060e23253e7e7763ac882b07c8f54b113a3