Fix errors in IntMap desktop implementation
The Desktop (JVM) implementation of IntMap has a few trivial errors
that cause it to behave incorrectly in all but the simplest cases. This
CL fixes those errors and fleshes out the test suite a bit.

This CL also corrects the expect/actual declaration of IntMap to be
consistent across Android and Desktop, which lets us share test suites
between the two implementations.

Test: IntMapTest
Change-Id: Icd814affc1f295dc1b952866fc73403bb54a5a23