Expose PlatformLocale and add hintLocales to ImeOptions
This is the first step of adding ime hint locales to TextField API in foundation. We need to expose the PlatformLocale object from ui before adding a logic that depends on it in foundation.

`PlatformLocale` is converted to an expect/actual class to prevent consumers from extending it and providing their own implementation.

ImeOptions take a nullable LocaleList as hintLocales to define a default state where hint locales are not specified at all. A lint error associated with nullable collecting in API surface is explicitly suppressed.

Test: :compose:ui:ui-text:testDebug
Test: :compose:ui:ui-text:cAT
Relnote: "Added `hintLocales` to `ImeOptions` to provide TextFields with the ability to hint IMEs with specific locales to preset a preferred language."
Relnote: "Expose `platformLocale` property from `Locale` that returns the underlying platform object, e.g. `java.util.Locale`."
Change-Id: I921c6c846078caddeed0cb05a027b5ae91b90c75