Refactor publication and artifacts redirecting configuration (#1180)
With this PR it's much easier to add new modules with redirecting
publication.
I use this approach in my branch for lifecycle publication. But want to
merge it to jb-main first.

Now we have our dedicated `JetbrainsAndroidXPlugin` which configures
`artifacts redirection` when applied (requires that Kotlin MPP plugin is
applied too).