Fix CfW tests in ui:ui module (#1303)
The tests didn't pass due to different issues.

- `CanvasBasedWindowTests.testPreventDefault` was failing due to missing
`key` value in the native KeyboardEvent
- `ComposeWindowLifecycleTest.allEvents` didn't work properly due to
browser policies:
https://developer.mozilla.org/en-US/docs/Web/API/Window/blur (the window
can't be blured from code). We open a new tab now to force `blur` evnet.
- MouseEventsTest: `runTest` was added to run the tests in TestScope. No
logical reason to do so now. But it's needed as a workaround for some
weird tests interference.

The tests should be run with an extra property:

`./gradlew :compose:ui:ui:wasmJsBrowserTest
-Pjetbrains.cfw.tests.useChrome=true`